Finishing a basic version of the Grid iterator tests.
Should be sufficient for smaller applications. There is at least one
memory leak in the Grid iterator. It's probably related to the creation of
new iterators by the begin and end functions, but I can't find any way
around that. This requires a bit more research and possibly some external
code review. Hopefully I figure it out eventually.

Right now, it's time to move on to the driver since I have a functional
core. I'll should come back to this later...